der: refactor `Int*`/`Uint*` types
Factors these types out of the (now-removed) `bigint` module into the
`int` and `uint` modules respectively, avoiding other code changes and
preserving all tests.
