+++ der/src/asn1.rs
@@ -35,7 +35,7 @@ pub use self::{
     context_specific::{ContextSpecific, ContextSpecificRef},
     generalized_time::GeneralizedTime,
     ia5_string::Ia5StringRef,
-    integer::bigint::{IntRef, UintRef},
+    integer::{int::IntRef, uint::UintRef},
     null::Null,
     octet_string::OctetStringRef,
     printable_string::PrintableStringRef,
@@ -54,7 +54,7 @@ pub use self::{
     any::Any,
     bit_string::BitString,
     ia5_string::Ia5String,
-    integer::bigint::{Int, Uint},
+    integer::{int::Int, uint::Uint},
     octet_string::OctetString,
     printable_string::PrintableString,
     set_of::SetOfVec,
